refactor: Centralize pagination logic on BaseTypeSafeCriteriaBuilder
